% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{accept_progress_bar}
\alias{accept_progress_bar}
\title{An altered version of the utils:txtProgressBar that shows acceptance rate}
\usage{
accept_progress_bar(min = 0, max = 1)
}
\arguments{
\item{min}{The minimum of the value being updated for the progress bar}

\item{max}{The maximum of the value being updated for the progress bar}
}
\value{
A structure matching the structure of a txtProgresBar with additional
  info
}
\description{
An altered version of the utils:txtProgressBar that shows acceptance rate
}
\keyword{internal}
