% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{accept_rate}
\alias{accept_rate}
\title{Return the acceptance rate for all subjects}
\usage{
accept_rate(store)
}
\arguments{
\item{store}{The samples store (containing random effects) with which we are
working}
}
\value{
A vector with the acceptance rate for each subject
}
\description{
Return the acceptance rate for all subjects
}
\keyword{internal}
