% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{acceptProgressBar}
\alias{acceptProgressBar}
\title{An altered version of the utils:txtProgressBar that shows acceptance rate}
\usage{
acceptProgressBar(min = 0, max = 1, initial = 0, char = "=", width = NA)
}
\arguments{
\item{min}{The minimum of the value being updated for the progress bar}

\item{max}{The maximum of the value being updated for the progress bar}

\item{initial}{The starting value for the progress bar}

\item{char}{The character to use as the indicator of progress}

\item{width}{The overall width of the progress bar}
}
\value{
A boolean TRUE or FALSE depending on the result of the test
}
\description{
An altered version of the utils:txtProgressBar that shows acceptance rate
}
\examples{
# No example yet
}
