% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{as_mcmc}
\alias{as_mcmc}
\title{Return a CODA mcmc object with the required samples}
\usage{
as_mcmc(sampler, selection = "theta_mu")
}
\arguments{
\item{sampler}{The pmwgs object containing samples to extract.}

\item{selection}{The selection of samples to return}
}
\value{
An mcmc object or list containing the selected samples.
}
\description{
Given a sampler object and a specification of the samples required, return
either an individual coda mcmc object, or a list of mcmc objects.
}
\examples{
# No example yet
}
