% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{augment_sampler_epsilon}
\alias{augment_sampler_epsilon}
\title{Augment existing sampler object to have subject specific epsilon storage}
\usage{
augment_sampler_epsilon(sampler)
}
\arguments{
\item{sampler}{The sampler object to augment}
}
\value{
A pmwgs sampler with epsilon array set internally
}
\description{
Older sampler object will be missing subject specific scaling parameter
(epsilon) storage, and running a stage with an updated pmwg will fail. To
fix this you can run the augment_sampler_epsilon function to fill the
appropriate array internals with NA values
}
