% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient.R
\name{check_adapted}
\alias{check_adapted}
\title{Check whether the adaptation phase has successfully completed}
\usage{
check_adapted(samples, unq_vals = 20)
}
\arguments{
\item{samples}{The subject mean samples with which we are working}

\item{unq_vals}{The number of unique values for each subject}
}
\value{
A boolean TRUE or FALSE depending on the result of the test
}
\description{
Check whether the adaptation phase has successfully completed
}
\keyword{internal}
