% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_efficient}
\alias{check_efficient}
\title{Check for efficient proposals if necessary}
\usage{
check_efficient(mix_ratio, efficient_mu, efficient_sig2)
}
\arguments{
\item{mix_ratio}{A vector of floats betwen 0 and 1 and summing to 1 which
give the ratio of particles to generate from the population level
parameters, the individual random effects and the conditional parameters
repectively}

\item{efficient_mu}{The mu value for the efficient proposals}

\item{efficient_sig2}{The sigma value for the efficient proposals}
}
\value{
nothing, stops operation on incorrect combiation of parameters.
}
\description{
Takes a mix ratio vector (3 x float) and the efficient proposal mu and sigma
If efficient proposals are to be used (mix_ratio[3] > 0) then test the
efficient proposal values to see whether they are not null and appropriate.
}
\examples{
pmwg:::check_efficient(c(0.1, 0.9, 0.0), NULL, NULL)
}
\keyword{internal}
