% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{check_run_stage_args}
\alias{check_run_stage_args}
\title{Test the arguments to the run_stage function for correctness}
\usage{
check_run_stage_args(
  pmwgs,
  stage,
  iter = 1000,
  particles = 1000,
  display_progress = TRUE,
  n_cores = 1,
  ...
)
}
\arguments{
\item{pmwgs}{A Particle Metropolis within Gibbs sampler which has been set
up and initialised}

\item{stage}{The sampling stage to run. Must be one of 'burn', 'adapt' or
'sample'. If not provided assumes that the stage should be 'burn'}

\item{iter}{The number of iterations to run for the sampler. For 'burn' and
'sample' all iteration will run. However for 'adapt' if all subjects have
enough unique samples to create the conditional distribution then it will
finish early.}

\item{particles}{The default here is 1000 particles to be generated for each
iteration, however during the sample phase this should be reduced.}

\item{display_progress}{Display a progress bar during sampling.}

\item{n_cores}{Set to > 1 to use mclapply}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Takes the arguments to run_stage and checks them for completeness and
correctness. Uses parent.frame to edit run_stage env directly
}
\keyword{internal}
