% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conditional_parms}
\alias{conditional_parms}
\title{Obtain the efficent mu and sigma from the adaptation phase draws}
\usage{
conditional_parms(samples, s)
}
\arguments{
\item{samples}{A list containing previous samples}

\item{s}{current subject number}
}
\value{
A list containing the conditional mean and variances for this subject
}
\description{
Obtain the efficent mu and sigma from the adaptation phase draws
}
\examples{
# No example yet
}
