% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient.R
\name{create_efficient}
\alias{create_efficient}
\title{Create distribution parameters for efficient proposals}
\usage{
create_efficient(x)
}
\arguments{
\item{x}{The current pmwgs object}
}
\value{
A list containing the mu and sigma for the proposal distribution.
}
\description{
From the existing samples, create a proposal distribution for drawing
efficient samples from.
}
\keyword{internal}
