% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract relevant samples from the list for conditional dist calc}
\usage{
extract_samples(samples)
}
\arguments{
\item{samples}{The samples list containing all samples from the pmwgs object}
}
\value{
A list containing only appopriate samples (non init/burnin samples)
}
\description{
From the existing samples, extract relevant samples for the creation of
the proposal distribution.
}
\examples{
# No example yet
}
\keyword{internal}
