% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forstmann}
\alias{forstmann}
\title{Forstmann et al.'s data.}
\format{A data frame with 15818 rows and 4 variables:
\describe{
  \item{subject}{subject, integer ID for each subject}
  \item{rt}{rt, reaction time for each trial as a double}
  \item{correct}{correct, integer code for correct or incorrect response}
  \item{condition}{condition, integer code for speed or accuracy condition}
}}
\source{
\url{https://www.jneurosci.org/content/35/6/2476}
}
\usage{
forstmann
}
\description{
A dataset containing the speed or accuracy manipulation, subject ID's and
reaction times from the ...
}
\keyword{datasets}
