% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{gen_proposals}
\alias{gen_proposals}
\title{Generate proposal particles.}
\usage{
gen_proposals(num_proposals, mu, sig2, particle, mix_ratio = 0.5)
}
\arguments{
\item{num_proposals}{A number representing the number of proposal particles to generate}

\item{mu}{A vector of means for the multivariate normal}

\item{sig2}{A covariate matrix for the multivariate normal}

\item{particle}{A particle (re proposals for latent variables)}

\item{mix_ratio}{A float betwen 0 and 1 giving the ratio of particles to generate from the population level parameters vs the individual level parameters}
}
\value{
The new proposals
}
\description{
Generate particles for a particular subject from a mix of either the
population level (hierarchical) distribution or from the particles
(containing individual level distribution).
}
\examples{
gen_proposals(100, rep(0.2, 7), diag(rep(0.1, 7)), rep(0.3, 7))
}
\keyword{internal}
