% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{gen_sample_pars}
\alias{gen_sample_pars}
\title{Initialise variables needed for individual loops within PMwG}
\usage{
gen_sample_pars(init, pmwg_args, prior, particles)
}
\arguments{
\item{init}{The list of parameter names to be used in the model}

\item{pmwg_args}{A list containing arguments to the PMwG sampling, including
number of iterations for each stage.}

\item{prior}{A list containing priors for variance, means etc.}

\item{particles}{A multidimensional array containing accepted particles from
the sampling.}
}
\value{
A list of generated variables that can be modified after the fact
}
\description{
Takes the init list and further sets up individual loop variables
}
\examples{
sampler <- pmwgs(forstmann, c("b1", "b2", "b3", "A", "v1", "v2", "t0"))
}
