% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{gibbs_step}
\alias{gibbs_step}
\title{Gibbs step of the Particle Metropolis within Gibbs sampelr}
\usage{
gibbs_step(samples, sampler)
}
\arguments{
\item{samples}{The list containing the samples from the current run, or from
the main storage in the sampler}

\item{sampler}{The pmwgs object from which to generate the new group
parameters.}
}
\value{
A list of generated variables that can be modified after the fact
}
\description{
Samples new theta_mu and theta_sig using Gibbs sampling
}
\examples{
# No example yet
}
\keyword{internal}
