% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{init}
\alias{init}
\title{Initialise a sampler with starting points}
\usage{
init(x, ...)
}
\arguments{
\item{x}{The sampling object that provides the parameters.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The sampler object but with initial values set for all parameter
  types
}
\description{
Takes the starting values for the individual and possibly group
level parameters, in preparation for further iterations,
whether they be burn-in or sampling.
}
\examples{
sampler <- pmwgs(forstmann, c("b1", "b2", "b3", "A", "v1", "v2", "t0"))
sampler <- init(sampler)
}
