% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_pmwg}
\alias{init_pmwg}
\title{Initialise variables needed for PMwG with default values}
\usage{
init_pmwg(parameters, data, pmwg_args)
}
\arguments{
\item{parameters}{The list of parameter names to be used in the model}

\item{data}{The data.frame containing empirical data to be modelled. Assumed to contain
at least one column called subject whose elements are unique identifiers for each subject.}

\item{pmwg_args}{A list containing arguments to the PMwG sampling, including number of
iterations for each stage.}
}
\value{
A list of generated variables that can be modified after the fact
}
\description{
Takes the initial parameter set and creates default values for a
range of PMwG necessary variables.
}
\examples{
# Load Forstmann et al.'s data.
args <- list(
  "burn_particles" = 1000,
  "burn_iter" = 500,
  "adapt_particles" = 100,
  "adapt_maxiter" = 5000,
  "sample_particles" = 100,
  "sample_iter" = 1000,
  "likelihood_func" = lba_loglike
)
psamplers:::init_pmwg(c("b1", "b2", "b3", "A", "v1", "v2", "t0"), forstmann, args)
}
