% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_sampler.R
\name{is.pmwgs}
\alias{is.pmwgs}
\title{Test whether object is pmwgs}
\usage{
is.pmwgs(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
logical, whether object inherits from pmwgs
}
\description{
Checks whether object is a Particle Metrolpolis with Gibbs sampler
}
