% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{last_sample}
\alias{last_sample}
\title{Create a list with the last samples in the pmwgs object}
\usage{
last_sample(store)
}
\arguments{
\item{store}{The list containing samples from t=which to grab the last.}
}
\value{
A list containing the last sample of group mean and variance and
  subject means.
}
\description{
Create a list with the last samples in the pmwgs object
}
\examples{
# No example yet
}
\keyword{internal}
