% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{lba_loglike}
\alias{lba_loglike}
\title{Function to calculate the log-likelihood of data, given random
effects, or coversely to produce synthetic data from given random
effects which match shape of "data". It relies on the rtdists package.}
\usage{
lba_loglike(x, data, sample = FALSE)
}
\arguments{
\item{x}{A list of parameter values}

\item{data}{A data.frame containing variables for
response time (rt), trial condition (condition)
and accuracy (correct) for which the likelihood
should be calculated.}
}
\value{
The log likelihood of \code{x} given \code{data}.
}
\description{
Function to calculate the log-likelihood of data, given random
effects, or coversely to produce synthetic data from given random
effects which match shape of "data". It relies on the rtdists package.
}
\examples{
x <- c(.2, .2, .2, .4, .3, 1.3, -2)
names(x) <- c("b1", "b2", "b3", "A", "v1", "v2", "t0")
rt <- c(0.432, 0.501, 0.31, 0.481, 0.342)
correct <- c(1, 2, 1, 1, 2)
condition <- c(1, 2, 3, 2, 1)
df <- data.frame(rt, correct, condition)
lba_loglike(x, df)
}
