% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{new_group_pars}
\alias{new_group_pars}
\title{Initialise variables needed for individual loops within PMwG}
\usage{
new_group_pars(samples, sampler)
}
\arguments{
\item{samples}{The list containing the samples from the current run, or from
the master storage in the sampler}

\item{sampler}{The pmwgs object from which to generate the new group
parameters.}
}
\value{
A list of generated variables that can be modified after the fact
}
\description{
Takes the pmwgs object and sets up sampling loop variables
}
\examples{
# No example yet
}
\keyword{internal}
