% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{new_group_pars_err}
\alias{new_group_pars_err}
\title{Error handler for the new_group_pars call}
\usage{
new_group_pars_err(pmwgs, store)
}
\arguments{
\item{err_cond}{The samples store (containing random effects) with which we are
working}
}
\value{
A vector with the acceptance rate for each subject
}
\description{
Error handler for the new_group_pars call
}
\examples{
# No example yet
}
\keyword{internal}
