% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMwG.R
\name{new_sample}
\alias{new_sample}
\title{Paralellisable particle sampling function (cond/orig/real/log).}
\usage{
new_sample(s, data, num_proposals, mu, sig2, particles, mix_ratio = 0.5)
}
\arguments{
\item{s}{A number - the subject ID, also selects particles}

\item{data}{A data.frame containing variables for
response time (\code{rt}), trial condition (\code{condition}),
accuracy (\code{correct}) and subject (\code{subject}) which
contains the data against which the particles are assessed}

\item{num_proposals}{A number representing the number of proposal particles to generate}

\item{mu}{A vector of means for the multivariate normal}

\item{sig2}{A covariate matrix for the multivariate normal}

\item{particles}{An array of particles (re proposals for latent variables)}
}
\value{
A single sample from the new proposals
}
\description{
Generate samples from either the initial proposal or from the last
iteration of the particles (according to a mixing value) and return
a weighted random sample from the new proposals
This uses the simplest, and slowest, proposals: a mixture of the
the population distribution and Gaussian around current random effect.
}
\examples{
# No example yet
}
