% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{numbers_from_proportion}
\alias{numbers_from_proportion}
\title{Check and normalise the number of each particle type from the mix_proportion}
\usage{
numbers_from_proportion(mix_proportion, num_particles = 1000)
}
\arguments{
\item{mix_proportion}{A vector of floats between 0 and 1 and summing to 1
which give the proportion of particles to generate from the population
level parameters, the individual random effects and the conditional
parameters respectively}

\item{num_particles}{The total number of particles to generate using a
combination of the three methods.}
}
\value{
The wound vector as a matrix
}
\description{
Takes a mix proportion vector (3 x float) and a number of particles to
generate and returns a vector containing the number of each particle type to
generate.
}
\keyword{internal}
