% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{particle_select_err}
\alias{particle_select_err}
\title{Error handler for the particle selection call}
\usage{
particle_select_err(subj, envir, err_cond)
}
\arguments{
\item{err_cond}{The original error condition that prompted this.}

\item{pmwgs}{The pmwgs object for the current run.}
}
\description{
If an error was detected when selecting the winning particle, save the state
of the samples and particles at that moment to help with debugging.
}
\keyword{internal}
