% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmwg-package.R
\docType{package}
\name{pmwg-package}
\alias{pmwg}
\alias{pmwg-package}
\title{pmwg: Particle Metropolis within Gibbs Sampler}
\description{
Provides an R implementation of the Particle Metropolis within
    Gibbs sampler for model parameter and random effect estimation. Based on
    the paper at https://doi.org/10.1016/j.jmp.2020.102368 and on the Matlab
    code available at http://osf.io/5b4w3 written by David Gunawan. A bookdown
	document describing the package is available at
	https://newcastlecl.github.io/samplerDoc/ and includes several detailed
	examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/newcastlecl/pmwg}
  \item Report bugs at \url{https://github.com/newcastlecl/pmwg/issues}
}

}
\author{
\strong{Maintainer}: Gavin Cooper \email{gavin@gavincooper.net} (Package creator, additional translation of MATLAB code) [translator]

Authors:
\itemize{
  \item Reilly Innes \email{Reilly.Innes@uon.edu.au}
  \item Caroline Kuhne \email{caroline.kuhne@newcastle.edu.au}
  \item Jon-Paul Cavallaro \email{jon-paul.cavallaro@uon.edu.au}
  \item David Gunawan \email{dgunawan@uow.edu.au} (Author of original MATLAB code from which this package was adapted)
  \item Guy Hawkins \email{guy.hawkins@newcastle.edu.au}
  \item Scott Brown \email{scott.brown@newcastle.edu.au} (Original translator from MATLAB to R) [translator]
}

}
\keyword{internal}
