% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmwg-package.R
\docType{package}
\name{pmwg}
\alias{pmwg}
\alias{pmwg-package}
\title{The pmwg Package}
\description{
Particle Sampling Methods
}
\details{
\tabular{ll}{
Package: \tab pmwg\cr
Type: \tab Package\cr
Version: \tab 0.0.0.9004\cr
Date: \tab 2019-07-23\cr
Depends: \tab R (>= 3.5.0)\cr
License: \tab GPL-3\cr
URL: \tab https://github.com/newcastlecl/pmwg\cr
}

Provides an R implementation of the Particle Metropolis within Gibbs sampler
for model parameter and random effect estimation. Based on the paper at
\url{https://doi.org/10.1016/j.jmp.2020.102368} and on the Matlab code
available at \url{http://osf.io/5b4w3}
}
\author{
Gavin Cooper
}
\keyword{package}
