% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psamplers.R
\docType{package}
\name{psamplers}
\alias{psamplers}
\alias{psamplers-package}
\title{Implementations of particle based sampling methods for model parameter estimation. Initially primarily an implementation of the Particle Metropolis within Gibbs sampler outlined in the paper available at https://arxiv.org/abs/1806.10089}
\description{
Implementations of particle based sampling methods for model parameter estimation. Initially primarily an implementation of the Particle Metropolis within Gibbs sampler outlined in the paper available at https://arxiv.org/abs/1806.10089
}
