% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendored.R
\name{riwish}
\alias{riwish}
\title{The Inverse Wishart Distribution}
\usage{
riwish(v, S)
}
\arguments{
\item{v}{Degrees of freedom (scalar).}

\item{S}{Scale matrix \eqn{(p \times p)}.}
}
\value{
\code{riwish} generates one random draw from the distribution.
}
\description{
Random generation from the Inverse Wishart distribution.
}
\details{
The mean of an inverse Wishart random variable with \code{v} degrees of
freedom and scale matrix \code{S} is \eqn{(v-p-1)^{-1}S}.
}
\examples{
\dontrun{
draw <- riwish(3, matrix(c(1,.3,.3,1),2,2))
}
}
\keyword{internal}
