% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{run_stage.pmwgs}
\alias{run_stage.pmwgs}
\title{Run a stage of the PMwG sampler}
\usage{
\method{run_stage}{pmwgs}(x, stage, iter = 1000, particles = 1000,
  display_progress = TRUE, n_cores = 1, ...)
}
\arguments{
\item{x}{A pmwgs object that has been initialised}

\item{stage}{The sampling stage to run. Must be one of 'burn', 'adapt' or
'sample'. If not provided assumes that the stage should be 'burn'}

\item{iter}{The number of iterations to run for the sampler. For 'burn' and
'sample' all iteration will run. However for 'adapt' if all subjects have
enough unique samples to create the conditional distribution then it will
finish early.}

\item{particles}{The default here is 1000 particles to be generated for each
iteration, however during the sample phase this should be reduced.}

\item{display_progress}{Display a progress bar during sampling.}

\item{n_cores}{Set to > 1 to use mclapply}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A pmwgs object with the newly generated samples in place.
}
\description{
Run one of burnin, adaptation of sampling phase from the PMwG
sampler. Each stage involves slightly different processes, so for the
full PMwG we need to run this three times.
}
\examples{
# No example yet
}
