% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{run_stage}
\alias{run_stage}
\title{Initialise a sampler with starting points}
\usage{
run_stage(x, ...)
}
\arguments{
\item{x}{The sampling object that provides the parameters.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The sampler object but with sampled values for subject and group
  parameters. Exact sampling method depends on the stage being run.
}
\description{
Takes the starting values for the individual and possibly group
level parameters, in preparation for further iterations,
whether they be burn-in or sampling.
}
\examples{
# No example yet
}
