% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendored.R
\name{rwish}
\alias{rwish}
\alias{dwish}
\title{The Wishart Distribution}
\usage{
rwish(v, S)
}
\arguments{
\item{v}{Degrees of freedom (scalar).}

\item{S}{Inverse scale matrix \eqn{(p \times p)}.}
}
\value{
\code{rwish} generates one random draw from the distribution.
}
\description{
Random generation from the Wishart distribution.
}
\details{
The mean of a Wishart random variable with \code{v} degrees of freedom and
inverse scale matrix \code{S} is \eqn{vS}.
}
\examples{
\dontrun{
draw <- rwish(3, matrix(c(1,.3,.3,1),2,2))
}
}
\keyword{internal}
