% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sample_store}
\alias{sample_store}
\title{Create a new list for storage samples in the pmwgs object}
\usage{
sample_store(par_names, n_subjects, iters = 1, stage = "init")
}
\arguments{
\item{par_names}{The names of each parameter as a character vector}

\item{n_subjects}{The number of subjects for the subject mean storage.}

\item{iters}{The number of iterations to be pre-allocated}

\item{stage}{The stage for which the samples will be created. Should be one
of `c("init", "burn", "adapt", "sample")`}
}
\value{
A list containing the conditional mean and variances for this subject
}
\description{
Create a new list for storage samples in the pmwgs object
}
\examples{
# No example yet
}
