% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampled_forstmann}
\alias{sampled_forstmann}
\title{A sampled object of a model of the Forstmann dataset}
\format{
A pmwgs object minus the data. A pmwgs opbject is a list with a
specific structure and elements, as outlined below.
\describe{
  \item{par_names}{A character vector containing the model parameter names}
  \item{n_pars}{The number of parameters in the model}
  \item{n_subjects}{The number of unique subject ID's in the data}
  \item{subjects}{A vector containing the unique subject ID's}
  \item{prior}{A list that holds the prior for \code{theta_mu} (the model
    parameters). Contains the mean (\code{theta_mu_mean}), covariance matrix
    (\code{theta_mu_var}) and inverse covariance matrix
    (\code{theta_mu_invar})}
  \item{ll_func}{The log likielihood function used by pmwg for model
    estimation}
  \item{samples}{A list with defined structure containing the samples, see
    the Samples Element section for more detail}
}
}
\source{
\url{https://www.pnas.org/content/105/45/17538}
}
\usage{
sampled_forstmann
}
\description{
A pmwgs object with a limited number of samples of the Forstmann dataset.
}
\details{
The pmwgs object is missing one aspect, the pmwgs$data element. In order
to fully replicate the full object (ie to run more sampling stages) you will
need to add the data back in, via sampled_forstmann$data <- forstmann
}
\section{}{
NA
}

\keyword{datasets}
