% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{set_epsilon}
\alias{set_epsilon}
\title{Set default values for epsilon}
\usage{
set_epsilon(n_pars, epsilon)
}
\arguments{
\item{n_pars}{The number of parameters for the model}

\item{epsilon}{The value of epsilon set in the call to `run_stage` or NULL.}
}
\description{
Takes the number of parameters and the epsilon arg value and sets a default
if necessary.
}
\keyword{internal}
