% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{set_mix}
\alias{set_mix}
\title{Set default values for mix}
\usage{
set_mix(stage, mix)
}
\arguments{
\item{stage}{The stage being run by the sampler.}

\item{mix}{The value of mix set in the call to `run_stage` or NULL.}
}
\description{
Takes the current stage and the mix arg value and sets a default if
necessary.
}
\keyword{internal}
