% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{set_proposal}
\alias{set_proposal}
\title{Setup the proposal distribution arguments (if in sample stage)}
\usage{
set_proposal(stage, pmwgs)
}
\arguments{
\item{stage}{The stage being run by the sampler.}

\item{pmwgs}{The pmwgs object from which to attempt to create the proposal
distribution}
}
\description{
Takes the current stage and the mix arg value and sets a default if
necessary.
}
\keyword{internal}
