% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient.R
\name{test_sampler_adapted}
\alias{test_sampler_adapted}
\title{Test that the sampler has successfully adapted}
\usage{
test_sampler_adapted(stage_samples, pmwgs, n_unique, i)
}
\arguments{
\item{stage_samples}{The samples store for the currently running stage.}

\item{pmwgs}{The full pmwgs object with all samples from previous runs.}

\item{n_unique}{The number of unique samples to look for in random effects
for each participant.}

\item{i}{The number for the current iteration of the sampler}
}
\value{
A string representing successful/unsuccessful adaptation. Can be one
  of c("success", "continue", "increase")
}
\description{
Test that the sampler has successfully adapted
}
\keyword{internal}
