% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{trim_na}
\alias{trim_na}
\title{Trim the unneeded NA values from the end of the sampler}
\usage{
trim_na(sampler)
}
\arguments{
\item{sampler}{The pmwgs object that we are adding the new samples to}
}
\value{
The pmwgs object without NA values added during extend_sampler
}
\description{
Trim the unneeded NA values from the end of the sampler
}
\keyword{internal}
