% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unwind}
\alias{unwind}
\title{Unwinds variance matrix to a vector}
\usage{
unwind(var_matrix, ...)
}
\arguments{
\item{var_matrix}{A variance matrix}
}
\value{
The unwound matrix as a vector
}
\description{
Takes a variance matrix and unwind to a vector via Cholesky then log
}
\examples{
psamplers:::unwind(diag(rep(1, 7)))
}
\keyword{internal}
