% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{update_epsilon}
\alias{update_epsilon}
\title{Update the subject specific scaling parameters (epsilon)}
\usage{
update_epsilon(epsilon, acc, p, i, d, alpha)
}
\arguments{
\item{epsilon}{The scaling parameter for all subjects}

\item{acc}{A boolean vector, TRUE if current sample != last sample}

\item{p}{The target sample acceptance rate (0-1)}

\item{i}{The current iteration for sampling}

\item{d}{The number of parameters for the model}

\item{alpha}{A hyperparameter for the epsilon tuning}
}
\value{
A vector with the new subject specific epsilon values
}
\description{
Update the subject specific scaling parameter according to procedures
outlined in  P. H. Garthwaite, Y. Fan & S. A. Sisson (2016) Adaptive optimal
scaling of Metropolis–Hastings algorithms using the Robbins–Monro process,
Communications in Statistics - Theory and Methods, 45:17, 5098-5111,
DOI: 10.1080/03610926.2014.936562
}
\keyword{internal}
