% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{update_progress_bar}
\alias{update_progress_bar}
\title{A function that updates the accept_progress_bar with progress and accept rate}
\usage{
update_progress_bar(pb, value, extra = 0)
}
\arguments{
\item{pb}{The progress bar object}

\item{value}{The value to set the bar width to}

\item{extra}{A value that represents the number of accepted particles}
}
\value{
The old value that was present before updating
}
\description{
A function that updates the accept_progress_bar with progress and accept rate
}
\keyword{internal}
