% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{update_sampler}
\alias{update_sampler}
\title{Update the main data store with the results of the last stage}
\usage{
update_sampler(sampler, store)
}
\arguments{
\item{sampler}{The pmwgs object that we are adding the new samples to}

\item{store}{The sample storage stage just run}
}
\value{
The pmwgs object with the new samples concatenated to the old
}
\description{
Update the main data store with the results of the last stage
}
\keyword{internal}
