% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{conditional_parms}
\alias{conditional_parms}
\title{Obtain the efficent mu and sigma from the adaptation phase draws}
\usage{
conditional_parms(init, ptm, pts2, s, start_idx, end_idx)
}
\arguments{
\item{init}{list containing previous samples and more}

\item{ptm}{a single iteration version of the param_theta_mu variable}

\item{pts2}{a single iter version of the param_theta_sigma2 variable}

\item{s}{current subject number}

\item{start_idx}{index specifying the starting sample to draw conditional distribution from}

\item{end_idx}{index specifying the end sample to draw coditional distribution from}
}
\value{
A list containing the conditional mean and variances for this subject
}
\description{
Obtain the efficent mu and sigma from the adaptation phase draws
}
\examples{
# No example yet
}
