% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient.R
\name{conditional_parms}
\alias{conditional_parms}
\title{Obtain the efficent mu and sigma from the adaptation phase draws}
\usage{
conditional_parms(s, samples)
}
\arguments{
\item{s}{current subject number}

\item{samples}{A list containing previous samples}
}
\value{
A list containing the conditional mean and variances for this subject
}
\description{
Obtain the efficent mu and sigma from the adaptation phase draws
}
\keyword{internal}
