% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract relevant samples from the list for conditional dist calc}
\usage{
extract_samples(sampler, stage = c("adapt", "sample"))
}
\arguments{
\item{sampler}{The pmwgs object containing the samples}

\item{stage}{The stage, or list of stages from which you want the samples}
}
\value{
A list containing only appropriate samples (non init/burnin samples)
}
\description{
From the sampler, extract relevant samples for the creation of
the proposal distribution.
}
\keyword{internal}
