% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forstmann}
\alias{forstmann}
\title{Forstmann et al.'s data.}
\format{
A data frame with 15818 rows and 5 variables:
\describe{
  \item{subject}{subject, integer ID for each subject}
  \item{rt}{rt, reaction time for each trial as a double}
  \item{condition}{condition, Factor with 3 levels for Speed, Accuracy and Neutral}
  \item{stim}{stim, Factor with 2 levels for Left and Right trials}
  \item{resp}{resp, Factor with 2 levels for Left and Right responses}
}
}
\source{
\url{https://www.jneurosci.org/content/35/6/2476}
}
\usage{
forstmann
}
\description{
A dataset containing the speed or accuracy manipulation, subject ID's and
reaction times from the ...
}
\keyword{datasets}
