% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gibbs_step_err}
\alias{gibbs_step_err}
\title{Error handler for the gibbs_step call}
\usage{
gibbs_step_err(pmwgs, store)
}
\arguments{
\item{err_cond}{The samples store (containing random effects) with which we are
working}
}
\value{
A vector with the acceptance rate for each subject
}
\description{
Error handler for the gibbs_step call
}
\examples{
# No example yet
}
\keyword{internal}
