% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{gibbs_step_err}
\alias{gibbs_step_err}
\title{Error handler for the gibbs_step call}
\usage{
gibbs_step_err(pmwgs, err_cond)
}
\arguments{
\item{pmwgs}{The pmwgs object for the current run.}

\item{err_cond}{The original error condition that prompted this.}
}
\description{
If an error was detected when generating new values in Gibbs step this
function is called to generate the error message and save the state of the
samples at that moment to help with debugging.
}
\keyword{internal}
