% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{gibbs_step_err}
\alias{gibbs_step_err}
\title{Error handler for the gibbs_step call}
\usage{
gibbs_step_err(pmwgs, store)
}
\arguments{
\item{pmwgs}{The pmwgs object for the current run.}

\item{store}{The samples store (containing random effects) with which we
are working in the current stage.}
}
\description{
If an error was detected when generating new values in Gibbs step this
function is called to generate the error message and save the state of the
samples at that moment to help with debugging.
}
\keyword{internal}
