% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{particle_draws}
\alias{particle_draws}
\title{Generate a cloud of particles from a multivariate normal distribution}
\usage{
particle_draws(n, mu, covar)
}
\arguments{
\item{n}{number of observations}

\item{mu}{mean vector}

\item{covar}{covariance matrix}
}
\value{
If n > 0 returns n draws from the multivariate normal with mean and
  sigma, otherwise returns NULL
}
\description{
Takes the mean and variance for a multivariate normal distribution, as well
as the number of particles to generate and return random draws from the
multivariate normal if the numbers of particles is > 0, otherwise return
NULL. At least one of mean or sigma must be provided.
}
\keyword{internal}
