% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_sampler.R
\name{pmwgs}
\alias{pmwgs}
\title{Initialise a PMwG sampler and return a pmwgs object}
\usage{
pmwgs(data, pars, ll_func, prior = NULL, ...)
}
\arguments{
\item{data}{The data.frame containing empirical data to be modelled. Assumed
to contain at least one column called subject whose elements are unique
identifiers for each subject.}

\item{pars}{The list of parameter names to be used in the model}

\item{ll_func}{A log likelihood function that given a list of parameter
values and a data.frame (or other data store) containing subject data will
return the log likelihood of \code{x} given \code{data}.}

\item{prior}{Specification of the prior distribution for mu and sigma2}

\item{...}{Other tuning parameters that you want to specify for the}
}
\value{
A pmwgs object that can be run, plotted and more
}
\description{
Takes the initial parameter set and creates default values for a
range of PMwG necessary variables.
}
\examples{
# Specify the log likelihood function
lba_loglike <- function(x, data) {
  x <- exp(x)
  if (any(data$rt < x["t0"])) {
    return(-1e10)
  }
  # This is faster than "paste".
  bs <- x["A"] + x[c("b1", "b2", "b3")][data$condition]

  out <- rtdists::dLBA(rt = data$rt, # nolint
                       response = data$correct,
                       A = x["A"],
                       b = bs,
                       t0 = x["t0"],
                       mean_v = x[c("v1", "v2")],
                       sd_v = c(1, 1),
                       distribution = "norm",
                       silent = TRUE)
  bad <- (out < 1e-10) | (!is.finite(out))
  out[bad] <- 1e-10
  out <- sum(log(out))
  out
}

# Specify parameter names and priors
pars <- c("b1", "b2", "b3", "A", "v1", "v2", "t0")
priors <- list(
  group_mean = rep(0, length(pars)),
  group_var = diag(rep(1, length(pars)))
)

# Create the Particle Metropolis within Gibbs sampler object
sampler <- pmwgs(
  data = forstmann,
  parameters = pars,
  llfunc = lba_loglike,
  prior = priors
)
}
