% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_sampler.R
\name{pmwgs}
\alias{pmwgs}
\title{Initialise a PMwG sampler and return a pmwgs object}
\usage{
pmwgs(data, parameters, prior = NULL, llfunc = lba_loglike)
}
\arguments{
\item{data}{The data.frame containing empirical data to be modelled. Assumed
to contain at least one column called subject whose elements are unique
identifiers for each subject.}

\item{parameters}{The list of parameter names to be used in the model}

\item{prior}{Specification of the prior distribution for mu and sigma2}

\item{llfunc}{A log likelihood function that given a list of parameter values
and a data.frame containing subject data will return the log likelihood of
\code{x} given \code{data}. See \code{\link{lba_loglike}} for an example.}
}
\value{
A pmwgs object that can be run, plotted and more
}
\description{
Takes the initial parameter set and creates default values for a
range of PMwG necessary variables.
}
\examples{
sampler <- pmwgs(forstmann, c("b1", "b2", "b3", "A", "v1", "v2", "t0"))
}
