% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_sampler.R
\name{pmwgs}
\alias{pmwgs}
\title{Initialise a PMwG sampler and return a pmwgs object}
\usage{
pmwgs(data, parameters, burn = c(1000, 500), adapt = c(1000, 5000),
  sample = c(100, 1000), llfunc = lba_loglike)
}
\arguments{
\item{data}{The data.frame containing empirical data to be modelled. Assumed
to contain at least one column called subject whose elements are unique
identifiers for each subject.}

\item{parameters}{The list of parameter names to be used in the model}

\item{pmwg_args}{A list containing arguments to the PMwG sampling, including
number of iterations for each stage.}
}
\value{
A pmwgs object that can be run, plotted and more
}
\description{
Takes the initial parameter set and creates default values for a
range of PMwG necessary variables.
}
\examples{
sampler <- pmwgs(forstmann, c("b1", "b2", "b3", "A", "v1", "v2", "t0"))
}
