% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psamplers-package.R
\docType{package}
\name{psamplers}
\alias{psamplers}
\alias{psamplers-package}
\title{The psamplers Package}
\description{
Particle Sampling Methods
}
\details{
\tabular{ll}{
Package: \tab psamplers\cr
Type: \tab Package\cr
Version: \tab 0.0.0.9004\cr
Date: \tab 2019-07-23\cr
Depends: \tab R (>= 3.5.0)\cr
License: \tab GPL-3\cr
URL: \tab https://github.com/gjcooper/samplers\cr
}

Provides particle sampling helper functions (log-likelihood functions, particle sampling functions). The initial version of the code is for parameter estimation and is primarily an implementation of the Particle Metropolis within Gibbs sampler outlined in the paper available at https://arxiv.org/abs/1806.10089 and based on the Matlab code available at http://osf.io/5b4w3
}
\author{
Gavin Cooper
}
\keyword{package}
