% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{sample_store}
\alias{sample_store}
\title{Create a new list for storage samples in the pmwgs object}
\usage{
sample_store(par_names, subject_ids, iters = 1, stage = "init")
}
\arguments{
\item{par_names}{The names of each parameter as a character vector}

\item{iters}{The number of iterations to be pre-allocated}

\item{stage}{The stage for which the samples will be created. Should be one
of \code{c("init", "burn", "adapt", "sample")}}

\item{subjects_ids}{The unique ID of each subjects as a character vector}
}
\value{
A list containing the conditional mean and variances for this subject
}
\description{
Create a new list for storage samples in the pmwgs object
}
\keyword{internal}
