% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient.R
\name{test_sampler_adapted}
\alias{test_sampler_adapted}
\title{Test that the sampler has successfully adapted}
\usage{
test_sampler_adapted(pmwgs, n_unique, i)
}
\arguments{
\item{pmwgs}{The full pmwgs object with all samples}

\item{n_unique}{The number of unique samples to look for in random effects
for each subject.}

\item{i}{The number for the current iteration of the sampler}
}
\value{
A list containing a string representing successful/unsuccessful
  adaptation and an optional message. The string representing the success
  or failure can be one of c("success", "continue", "increase")
}
\description{
Test that the sampler has successfully adapted
}
\keyword{internal}
