% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{wind}
\alias{wind}
\title{Winds a variance vector back to a vector}
\usage{
wind(var_vector, ...)
}
\arguments{
\item{var_vector}{A variance vector}
}
\value{
The wound vector as a matrix
}
\description{
The reverse of the unwind function, takes a variance vector and windows back
into matrix
}
\keyword{internal}
